+++ Cargo.toml
@@ -45,6 +45,7 @@ petgraph-algorithms = { path = "crates/algorithms", default-features = false }
 petgraph-generators = { path = "crates/generators", default-features = false }
 petgraph-proptest = { path = "crates/proptest", default-features = false }
 petgraph-io = { path = "crates/io", default-features = false }
+petgraph-test-utils = { path = "crates/test-utils", default-features = false }
 
 [dependencies]
 petgraph-adjacency-matrix = { workspace = true, optional = true, default-features = false }
