+++ crates/graph/Cargo.toml
@@ -21,6 +21,7 @@ serde_json = "1.0.96"
 
 criterion = "0.5.1"
 rand = "0.8.5"
+petgraph-test-utils.workspace = true
 
 [features]
 default = ["std", "stable"]
