+++ crates/matrix-graph/benches/common/mod.rs
@@ -1,0 +1,1 @@
+pub use petgraph_test_utils::*;
