+++ crates/test-utils/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "petgraph-test-utils"
+version = "0.0.0"
+edition = "2021"
+# INTERNAL CRATE FOR TESTING, DO NOT PUBLISH
+publish = false
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+petgraph.workspace = true
+rand = "0.8.5"
