+++ crates/test-utils/src/factories.rs
@@ -1,25 +1,31 @@
 use std::marker::PhantomData;
 
-use petgraph::{data::Build, visit::NodeIndexable, EdgeType};
-use petgraph_core::{
-    edge::{Directed, Undirected},
-    visit::GraphProp,
+use petgraph::{
+    core::{
+        edge::{Directed, Undirected},
+        visit::GraphProp,
+    },
+    data::Build,
+    graph::Graph,
+    visit::NodeIndexable,
+    EdgeType,
 };
-use petgraph_graph::Graph;
 
 /// Petersen A and B are isomorphic
 ///
 /// http://www.dharwadker.org/tevet/isomorphism/
-const PETERSEN_A: &str = include_str!("../../tests/snapshots/isomorphism/petersen_a.txt");
-const PETERSEN_B: &str = include_str!("../../tests/snapshots/isomorphism/petersen_b.txt");
+const PETERSEN_A: &str =
+    include_str!("../../algorithms/tests/snapshots/isomorphism/petersen_a.txt");
+const PETERSEN_B: &str =
+    include_str!("../../algorithms/tests/snapshots/isomorphism/petersen_b.txt");
 
 /// An almost full set, isomorphic
-const FULL_A: &str = include_str!("../../tests/snapshots/isomorphism/full_a.txt");
-const FULL_B: &str = include_str!("../../tests/snapshots/isomorphism/full_b.txt");
+const FULL_A: &str = include_str!("../../algorithms/tests/snapshots/isomorphism/full_a.txt");
+const FULL_B: &str = include_str!("../../algorithms/tests/snapshots/isomorphism/full_b.txt");
 
 /// Praust A and B are not isomorphic
-const PRAUST_A: &str = include_str!("../../tests/snapshots/isomorphism/praust_a.txt");
-const PRAUST_B: &str = include_str!("../../tests/snapshots/isomorphism/praust_b.txt");
+const PRAUST_A: &str = include_str!("../../algorithms/tests/snapshots/isomorphism/praust_a.txt");
+const PRAUST_B: &str = include_str!("../../algorithms/tests/snapshots/isomorphism/praust_b.txt");
 
 const BIGGER: &str = "
  0 0 0 0 0 0 0 0 1 0 1 0 0 0 0 0 0 0 1 1 0 1 1 1 1 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0
@@ -139,47 +145,58 @@ where
         }
     }
 
+    #[must_use]
     pub fn petersen_a(self) -> G {
         parse_graph(PETERSEN_A)
     }
 
+    #[must_use]
     pub fn petersen_b(self) -> G {
         parse_graph(PETERSEN_B)
     }
 
+    #[must_use]
     pub fn full_a(self) -> G {
         parse_graph(FULL_A)
     }
 
+    #[must_use]
     pub fn full_b(self) -> G {
         parse_graph(FULL_B)
     }
 
+    #[must_use]
     pub fn praust_a(self) -> G {
         parse_graph(PRAUST_A)
     }
 
+    #[must_use]
     pub fn praust_b(self) -> G {
         parse_graph(PRAUST_B)
     }
 
+    #[must_use]
     pub fn bigger(self) -> G {
         parse_graph(BIGGER)
     }
 
+    #[must_use]
     pub fn bipartite(self) -> G {
         parse_graph(BIPARTITE)
     }
 }
 
+#[must_use]
 pub const fn graph<Ty: EdgeType>() -> GraphFactory<Graph<(), (), Ty>> {
     GraphFactory::new()
 }
 
+#[must_use]
 pub const fn undirected_graph() -> GraphFactory<Graph<(), (), Undirected>> {
     graph()
 }
 
+#[must_use]
 pub const fn directed_graph() -> GraphFactory<Graph<(), (), Directed>> {
     graph()
 }
